+++ temba/channels/views.py
@@ -862,7 +862,7 @@ class ChannelCRUDL(SmartCRUDL):
     model = Channel
     actions = ('list', 'claim', 'update', 'read', 'delete', 'search_numbers', 'claim_twilio',
                'claim_android', 'claim_africas_talking', 'claim_chikka', 'configuration',
-               'search_nexmo', 'claim_nexmo', 'bulk_sender_options', 'create_bulk_sender', 'claim_infobip',
+               'search_nexmo', 'claim_nexmo', 'bulk_sender_options', 'create_bulk_sender',
                'claim_hub9', 'claim_vumi', 'claim_vumi_ussd', 'create_caller', 'claim_kannel', 'claim_shaqodoon',
                'claim_verboice', 'claim_clickatell', 'claim_plivo', 'search_plivo', 'claim_high_connection', 'claim_blackmyna',
                'claim_smscentral', 'claim_start', 'claim_m3tech', 'claim_yo', 'claim_viber', 'create_viber',
@@ -1573,10 +1573,6 @@ class ChannelCRUDL(SmartCRUDL):
 
             return super(ChannelCRUDL.ClaimAuthenticatedExternal, self).form_valid(form)
 
-    class ClaimInfobip(ClaimAuthenticatedExternal):
-        title = _("Connect Infobip")
-        channel_type = Channel.TYPE_INFOBIP
-
     class ClaimBlackmyna(ClaimAuthenticatedExternal):
         title = _("Connect Blackmyna")
         channel_type = Channel.TYPE_BLACKMYNA
@@ -1634,10 +1630,13 @@ class ChannelCRUDL(SmartCRUDL):
             password = forms.CharField(label=_("Password"),
                                        help_text=_("The password to be used to authenticate to Junebug"),
                                        required=False)
+            secret = forms.CharField(label=_("Secret"),
+                                     help_text=_("The token Junebug should use to authenticate"),
+                                     required=False)
 
         title = _("Connect Junebug")
         form_class = JunebugForm
-        fields = ('channel_type', 'country', 'number', 'url', 'username', 'password')
+        fields = ('channel_type', 'country', 'number', 'url', 'username', 'password', 'secret')
 
         def form_valid(self, form):
             org = self.request.user.get_org()
@@ -1654,6 +1653,9 @@ class ChannelCRUDL(SmartCRUDL):
                                                                      data['password'], data['channel_type'],
                                                                      data.get('url'),
                                                                      role=role)
+            if data['secret']:
+                self.object.secret = data['secret']
+                self.object.save()
 
             return super(ChannelCRUDL.ClaimAuthenticatedExternal, self).form_valid(form)